# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegistrationStatus": {
        "description": "Enumeration for the registration status of a customer.\n\nAttributes:\n    ACTIVE (str): The customer is registered and may be quoted and served.\n    STOPPED (str): The customer has stopped the service; no new quote or\n        intervention may be created for them.",
        "enum": [
          "active",
          "stopped"
        ],
        "title": "RegistrationStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Quoting and intervention planning for a home-care agency: customers, assistants, intervention types, quotes and the planning computation.",
    "title": "rt-erp API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/customers": {
      "get": {
        "description": "List customers.\n\nArgs:\n    page (int): One-based page number.\n    size (int): Page size.\n    search (Optional[str]): Case-insensitive fragment of a name.\n    registration_status (Optional[RegistrationStatus]): Restrict to one\n        status.\n    service (CustomerService): The customer service.\n    _ (User): The authenticated caller; enforces manager access.\n\nReturns:\n    List[Customer]: The matching customers.",
        "operationId": "list_customers_api_v1_customers_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RegistrationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Customer-Output"
                  },
                  "title": "Response List Customers Api V1 Customers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Customers",
        "tags": [
          "Customers"
        ]
      },
      "post": {
        "description": "Register a customer.\n\nArgs:\n    customer (Customer): The customer to register.\n    service (CustomerService): The customer service.\n    _ (User): The authenticated caller; enforces manager access.\n\nReturns:\n    Customer: The stored customer, with its identifier.\n\nNotes:\n    The address resolves its own coordinate while the payload is validated,\n    so this call may wait on the geocoding service. A street the map does\n    not know is still accepted; the failure is recorded on the address.",
        "operationId": "create_customer_api_v1_customers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Customer",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/customers/{customer_id}": {
      "delete": {
        "description": "Remove a customer who has never been quoted.\n\nArgs:\n    customer_id (str): The customer to remove.\n    service (CustomerService): The customer service.\n    _ (User): The authenticated caller; enforces manager access.\n\nRaises:\n    MTCustomerNotFound: If no such customer exists; answered as a 404.\n    MTCustomerHasQuotes: If any quote names them; answered as a 409, with\n        stopping them offered instead.",
        "operationId": "delete_customer_api_v1_customers__customer_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Customer",
        "tags": [
          "Customers"
        ]
      },
      "get": {
        "description": "Return one customer.\n\nArgs:\n    customer_id (str): The customer to read.\n    service (CustomerService): The customer service.\n    _ (User): The authenticated caller; enforces manager access.\n\nReturns:\n    Customer: The customer.\n\nRaises:\n    MTCustomerNotFound: If no such customer exists; answered as a 404.",
        "operationId": "get_customer_api_v1_customers__customer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Customer",
        "tags": [
          "Customers"
        ]
      },
      "put": {
        "description": "Replace a customer's details.\n\nArgs:\n    customer_id (str): The customer to change.\n    customer (Customer): The new details.\n    service (CustomerService): The customer service.\n    _ (User): The authenticated caller; enforces manager access.\n\nReturns:\n    Customer: The updated customer.\n\nRaises:\n    MTCustomerNotFound: If no such customer exists; answered as a 404.\n\nNotes:\n    The path identifier wins over anything in the body \u2014 a payload naming a\n    different customer edits the one that was addressed, not the one it\n    names.",
        "operationId": "update_customer_api_v1_customers__customer_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Customer",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/customers/{customer_id}/quotes": {
      "get": {
        "description": "Return every quote issued to a customer.\n\nArgs:\n    customer_id (str): The customer to read.\n    service (CustomerService): The customer service.\n    _ (User): The authenticated caller; enforces manager access.\n\nReturns:\n    List[Quote]: Their quotes.\n\nRaises:\n    MTCustomerNotFound: If no such customer exists; answered as a 404.",
        "operationId": "list_customer_quotes_api_v1_customers__customer_id__quotes_get",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Quote-Output"
                  },
                  "title": "Response List Customer Quotes Api V1 Customers  Customer Id  Quotes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Customer Quotes",
        "tags": [
          "Customers"
        ]
      }
    },
...
    "/api/v1/me/customers": {
      "get": {
        "description": "Return the customers the caller serves.\n\nArgs:\n    page (int): One-based page number.\n    size (int): Page size.\n    search (Optional[str]): Restrict by name or address.\n    service (CustomerService): The customer service.\n    caller (User): The authenticated caller.\n\nReturns:\n    List[Customer]: The caller's own customer portfolio.\n\nRaises:\n    MTHcaForbidden: If the account is bound to no assistant record;\n        answered as a 403.\n\nNotes:\n    The portfolio is the customers the assistant has a planned visit with,\n    plus those on quotes they wrote. It is **not** the agency's customer\n    directory: a home-care record carries an address, a telephone number and\n    a care schedule, and there is no reason for every assistant to hold every\n    one of them.",
        "operationId": "list_my_customers_api_v1_me_customers_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Customer-Output"
                  },
                  "title": "Response List My Customers Api V1 Me Customers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List My Customers",
        "tags": [
          "My account"
        ]
      }
    },
    "/api/v1/me/customers/{customer_id}": {
      "get": {
        "description": "Return one customer from the caller's portfolio.\n\nArgs:\n    customer_id (str): The customer to read.\n    service (CustomerService): The customer service.\n    caller (User): The authenticated caller.\n\nReturns:\n    Customer: The customer.\n\nRaises:\n    MTHcaForbidden: If the account is bound to no assistant record;\n        answered as a 403.\n    MTCustomerNotFound: If the customer does not exist **or** is not in the\n        caller's portfolio; answered as a 404 either way.\n\nNotes:\n    The same 404 whether the customer is absent or simply not theirs.\n    Distinguishing the two would let an assistant discover which identifiers\n    are real by trying them.",
        "operationId": "read_my_customer_api_v1_me_customers__customer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read My Customer",
        "tags": [
          "My account"
        ]
      }
    },
    "/api/v1/me/hca": {
      "get": {
        "description": "Return the caller's own assistant record.\n\nArgs:\n    service (HcaService): The assistant service.\n    caller (User): The authenticated caller.\n\nReturns:\n    HcaResponse: The caller's own record.\n\nRaises:\n    MTHcaForbidden: If the account is bound to no assistant record;\n        answered as a 403.\n    MTHcaNotFound: If the record has since been deleted; answered as a 404.",
        "operationId": "read_my_profile_api_v1_me_hca_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HcaResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Read My Profile",
        "tags": [
          "My account"
        ]
      },
      "patch": {
        "description": "Change the caller's own contact details and address.\n\nArgs:\n    request (HcaProfileUpdateRequest): The new details.\n    service (HcaService): The assistant service.\n    caller (User): The authenticated caller.\n\nReturns:\n    HcaResponse: The updated record.\n\nRaises:\n    MTHcaForbidden: If the account is bound to no assistant record;\n        answered as a 403.\n    MTHcaNotFound: If the record does not exist; answered as a 404.\n\nNotes:\n    **The contract type and the certifications cannot be changed here**, and\n    not because this endpoint ignores them \u2014 the payload has no such fields.\n    What an assistant is employed as, and what they are qualified to do, are\n    a manager's decisions, made through\n    ``PATCH /api/v1/hcas/{id}/employment``. An assistant who could grant\n    themselves a certification could be routed to work they are not trained\n    for.",
        "operationId": "update_my_profile_api_v1_me_hca_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HcaProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HcaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update My Profile",
        "tags": [
          "My account"
        ]
      }
    },
    "/api/v1/me/quotes": {
      "get": {
        "description": "Return the quotes the caller wrote.\n\nArgs:\n    page (int): One-based page number.\n    size (int): Page size.\n    service (QuoteService): The quote service.\n    caller (User): The authenticated caller.\n\nReturns:\n    List[Quote]: The caller's own quotes, whatever their status.",
        "operationId": "list_my_quotes_api_v1_me_quotes_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Quote-Output"
                  },
                  "title": "Response List My Quotes Api V1 Me Quotes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List My Quotes",
        "tags": [
          "My account"
        ]
      },
      "post": {
        "description": "Write a quote, as a draft the caller still owns.\n\nArgs:\n    quote (Quote): The quote to create.\n    service (QuoteService): The quote service.\n    caller (User): The authenticated caller.\n\nReturns:\n    Quote: The stored, priced quote.\n\nRaises:\n    MTPricingUnknownInterventionType: If a line names a type that is not in\n        the catalog; answered as a 422.\n\nNotes:\n    It is created as a **draft**, not submitted. Writing a quote and\n    deciding it is ready are two separate acts, and an assistant pricing up\n    a visit while sitting with a family should be able to save it and check\n    the figures before a manager is asked to look.",
        "operationId": "create_my_quote_api_v1_me_quotes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Quote-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create My Quote",
        "tags": [
          "My account"
        ]
      }
    },
    "/api/v1/me/quotes/{quote_id}/submit": {
      "post": {
        "description": "Send one of the caller's own drafts for validation.\n\nArgs:\n    quote_id (str): The quote to submit.\n    service (QuoteService): The quote service.\n    caller (User): The authenticated caller.\n\nReturns:\n    Quote: The submitted quote, now awaiting a manager.\n\nRaises:\n    MTQuoteNotFound: If no such quote exists; answered as a 404.\n    MTQuoteForbidden: If the caller did not write it; answered as a 403.\n    MTQuoteNotEditable: If it is not a draft; answered as a 409.\n    MTQuoteNotPriced: If it has no priced lines; answered as a 409.",
        "operationId": "submit_my_quote_api_v1_me_quotes__quote_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "quote_id",
            "required": true,
            "schema": {
              "title": "Quote Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit My Quote",
        "tags": [
          "My account"
        ]
      }
    },
...
  }
}